refactor: Update Makefile to start postgres server for tests
This commit modifies the Makefile to start a postgres server in the background and run the tests against it. The server will be stopped after the tests are done. This change improves the testing process by ensuring that the tests are run against a real database environment. The use of `$$` to escape the `$` character is necessary because the command is being run in a subshell. The DATABASE_URL environment variable is set using the `npx pg-test start` command. The Makefile also includes commands to migrate the database and run the tests using the `pnpm` package manager.