refactor: Add "biomejs.biome" to recommended extensions
The "biomejs.biome" extension has been added to the list of recommended extensions in the ".vscode/extensions.json" file. This change ensures that users are prompted to install the "biomejs.biome" extension when opening the project in Visual Studio Code. It improves the development experience by providing additional functionality specific to the Biome project.